+++ src/screens/App/screens/Seller/screens/Shipment/screens/Pack/components/table/PackOrderSkeleton.jsx
@@ -1,0 +1,44 @@
+import React from "react";
+import {
+  Data,
+  SkeletonForText,
+  SkeletonRow,
+} from "../../../../../../../../../../shared/Components/table/SpaciousTable";
+
+const PackOrderSkeleton = () => {
+  const dummy = [0, 0, 0, 0, 0];
+
+  return (
+    <>
+      {dummy.map((e, i) => (
+        <SkeletonRow key={i} grid="grid-cols-12">
+          <Data className="col-span-1">
+            <SkeletonForText />
+          </Data>
+          <Data className="col-span-1">
+            <SkeletonForText />
+          </Data>
+          <Data className="col-span-3">
+            <SkeletonForText />
+          </Data>
+          <Data className="col-span-2">
+            <SkeletonForText />
+          </Data>
+          <Data className="col-span-1">
+            <SkeletonForText />
+          </Data>
+          <Data className="col-span-1">
+            <SkeletonForText />
+          </Data>
+          <Data className="col-span-2">
+            <SkeletonForText />
+          </Data>
+          <Data className="col-span-1">
+            <SkeletonForText />
+          </Data>
+        </SkeletonRow>
+      ))}
+    </>
+  );
+};
+export default PackOrderSkeleton;
